requirements.txt
attrs==21.4.0
click==8.1.3
Flask==2.1.2
Flask-JWT==0.3.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==1.4.2
pyparsing==3.0.8
pytest==7.1.2
tomli==2.0.1
Werkzeug==2.1.2
zipp==3.8.0
